Data Base Connected and test some queries

+++ src/General.java
@@ -1,7 +1,7 @@
 import java.time.LocalDate;
 
 public class General extends Ticket {
-    public General(int ticketNumber, String category, String holder, String type, LocalDate date, int time) {
+    public General(int ticketNumber, String category, String holder, String type, LocalDate date, float time)  {
         super(ticketNumber, category, holder, type, date, time);
     }
 }
